refactor: move out common utils
