+++ otx/algorithms/detection/adapters/mmdet/nncf/monkey_patch.py
@@ -2,7 +2,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-from contextlib import contextmanager
 from functools import partial, partialmethod
 
 from mmdet import core
@@ -17,49 +16,17 @@ from mmdet.models.roi_heads.base_roi_head import BaseRoIHead
 from mmdet.models.roi_heads.mask_heads.fcn_mask_head import FCNMaskHead
 from mmdet.models.roi_heads.roi_extractors import SingleRoIExtractor
 
-from otx.algorithms.common.adapters.nncf.utils import (
-    is_nncf_enabled,
-    nncf_trace,
-    no_nncf_trace,
+from otx.algorithms.common.adapters.nncf.monkey_patch import (
+    conditioned_wrapper,
+    nncf_trace_context,
+    nncf_trace_wrapper,
+    no_nncf_trace_wrapper,
 )
 
 
 #  from nncf.torch.dynamic_graph.context import get_current_context
 #  if get_current_context() is not None and get_current_context().is_tracing:
 #      __import__('ipdb').set_trace()
-
-
-def no_nncf_trace_wrapper(*args, **kwargs):
-    """
-    A wrapper function not to trace in NNCF.
-    """
-
-    in_fn = kwargs.pop("in_fn")
-    with no_nncf_trace():
-        return in_fn(*args, **kwargs)
-
-
-def nncf_trace_wrapper(*args, **kwargs):
-    """
-    A wrapper function to trace in NNCF.
-    """
-
-    in_fn = kwargs.pop("in_fn")
-    with nncf_trace():
-        return in_fn(*args, **kwargs)
-
-
-def conditioned_wrapper(target_cls, wrapper, methods, subclasses=(object,)):
-    """
-    A function to wrap all the given methods under given subclasses.
-    """
-
-    if issubclass(target_cls, subclasses):
-        for func_name in methods:
-            func = getattr(target_cls, func_name, None)
-            if func is not None and "_partialmethod" not in func.__dict__:
-                #  print(target_cls, func_name)
-                setattr(target_cls, func_name, partialmethod(wrapper, in_fn=func))
 
 
 def wrap_mmdet_head(head_cls):
@@ -98,41 +65,6 @@ def wrap_register_module(self, *args, **kwargs):
     wrap_mmdet_bbox_assigner(module)
     wrap_mmdet_sampler(module)
     return in_fn(*args, **kwargs)
-
-
-@contextmanager
-def nncf_trace_context(self, img_metas):
-    """
-    A context manager for nncf graph tracing
-    """
-
-    assert (
-        getattr(self, "forward_backup", None) is None
-    ), "Error: one forward context inside another forward context"
-
-    if is_nncf_enabled():
-        from nncf.torch.nncf_network import NNCFNetwork
-
-        if isinstance(self, NNCFNetwork):
-            self.get_nncf_wrapped_model().forward_backup = self.forward
-
-    def forward_nncf_trace(self, img, img_metas, **kwargs):
-        return self.onnx_export(img[0], img_metas[0])
-        #  return self.simple_test(img[0], img_metas[0])
-
-    # onnx_export in mmdet head has a bug with cuda
-    self.device_backup = next(self.parameters()).device
-    self.forward_backup = self.forward
-    self = self.to("cpu")
-    self.forward = partialmethod(forward_nncf_trace, img_metas=img_metas).__get__(self)
-    yield
-    self = self.to(self.device_backup)
-    delattr(self, "device_backup")
-    self.forward = self.forward_backup
-    delattr(self, "forward_backup")
-
-    if is_nncf_enabled() and isinstance(self, NNCFNetwork):
-        self.get_nncf_wrapped_model().forward_backup = None
 
 
 # add nncf context method that will be used when nncf tracing
